requirements.txt
requests>=2.31.0
httpx>=0.24.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
pydantic>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0

# APIs de IA
openai>=1.0.0
voyageai>=0.2.0

# Banco de dados vetorial
astrapy>=1.0.0

# Processamento de documentos
llama-parse>=0.4.0